pyproject.toml
[project]
name = "deep-research-clone"
version = "0.1.0"
description = "A simple Deep Research clone with language models and web search"
authors = [
    {name = "Developer", email = "dev@example.com"}
]
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "markdown>=3.5.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "ollama>=0.1.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
deep-research = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
